Fix zoomed text cut-off
When HTML pages are zoomed in or the
window width is shortened, the text
is cut off on the left side because
the `body` is padded to the right by
pixels.

This PR resolves the issue by changing
the padding value to a percentage value
that looks similar to what we have now.
Since the padding is relative, the body
text will never be cut off.

We also need to break lines and words
to avoid 2-dimensional scrolling.